+++ lib/ui/src/chat/CodeBlocks.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useMemo, useRef } from 'react'
 
 import classNames from 'classnames'
 
-import { renderCodyMarkdown } from '@sourcegraph/cody-shared'
+import { Guardrails, renderCodyMarkdown } from '@sourcegraph/cody-shared'
 
 import { CodeBlockActionsProps } from '../Chat'
 import {
@@ -10,6 +10,7 @@ import {
     CopyCodeBlockIcon,
     InsertCodeBlockIcon,
     SaveCodeBlockIcon,
+    ShieldIcon,
 } from '../icons/CodeBlockActionIcons'
 
 import styles from './CodeBlocks.module.css'
@@ -26,6 +27,8 @@ interface CodeBlocksProps {
     insertButtonOnSubmit?: CodeBlockActionsProps['insertButtonOnSubmit']
 
     metadata?: CodeBlockMeta
+
+    guardrails?: Guardrails
 }
 
 export interface CodeBlockMeta {
@@ -159,6 +162,7 @@ export const CodeBlocks: React.FunctionComponent<CodeBlocksProps> = React.memo(f
     insertButtonOnSubmit,
     metadata,
     inProgress,
+    guardrails,
 }) {
     const rootRef = useRef<HTMLDivElement>(null)
 
@@ -177,7 +181,6 @@ export const CodeBlocks: React.FunctionComponent<CodeBlocksProps> = React.memo(f
             const preText = preElement.textContent
             if (preText?.trim() && preElement.parentNode) {
                 const eventMetadata = { requestID: metadata?.requestID, source: metadata?.source }
-
                 const buttons = createButtons(
                     preText,
                     copyButtonClassName,
@@ -186,6 +189,42 @@ export const CodeBlocks: React.FunctionComponent<CodeBlocksProps> = React.memo(f
                     insertButtonOnSubmit,
                     eventMetadata
                 )
+                if (guardrails) {
+                    const flexFiller = document.createElement('div')
+                    flexFiller.classList.add(styles.flexFiller)
+                    buttons.append(flexFiller)
+                    const attributionContainer = document.createElement('div')
+                    attributionContainer.innerHTML = ShieldIcon
+                    attributionContainer.classList.add(styles.attributionIcon)
+                    attributionContainer.title = 'Attribution search running...'
+                    buttons.append(attributionContainer)
+
+                    guardrails
+                        .searchAttribution(preText)
+                        .then(attribution => {
+                            if (attribution instanceof Error || attribution.limitHit) {
+                                attributionContainer.classList.add(styles.attributionIconUnavailable)
+                                attributionContainer.title = 'Attribution search unavailable.'
+                                return
+                            }
+                            if (attribution.repositories.length > 0) {
+                                attributionContainer.classList.add(styles.attributionIconFound)
+                                let tooltip = `Attribution found in ${attribution.repositories[0].name}`
+                                if (attribution.repositories.length > 1) {
+                                    tooltip = `${tooltip} and ${attribution.repositories.length - 1} more.`
+                                } else {
+                                    tooltip = `${tooltip}.`
+                                }
+                                attributionContainer.title = tooltip
+                                return
+                            }
+                            attributionContainer.classList.add(styles.attributionIconNotFound)
+                            attributionContainer.title = 'Attribution not found.'
+                        })
+                        .catch(error => {
+                            console.error('promise failed', error)
+                        })
+                }
 
                 // Insert the buttons after the pre using insertBefore() because there is no insertAfter()
                 preElement.parentNode.insertBefore(buttons, preElement.nextSibling)
@@ -208,6 +247,7 @@ export const CodeBlocks: React.FunctionComponent<CodeBlocksProps> = React.memo(f
         metadata?.requestID,
         metadata?.source,
         inProgress,
+        guardrails,
     ])
 
     return useMemo(