+++ vscode/src/chat/chat-view/SimpleChatPanelProvider.ts
@@ -6,6 +6,7 @@ import * as vscode from 'vscode'
 import { ActiveTextEditorSelectionRange, ChatMessage, ContextFile } from '@sourcegraph/cody-shared'
 import { ChatModelProvider } from '@sourcegraph/cody-shared/src/chat-models'
 import { ChatClient } from '@sourcegraph/cody-shared/src/chat/chat'
+import { isCodyIgnoredFile } from '@sourcegraph/cody-shared/src/chat/context-filter'
 import {
     createDisplayTextWithFileLinks,
     createDisplayTextWithFileSelection,
@@ -19,7 +20,6 @@ import { reformatBotMessageForChat } from '@sourcegraph/cody-shared/src/chat/vie
 import { ContextMessage } from '@sourcegraph/cody-shared/src/codebase-context/messages'
 import { Editor } from '@sourcegraph/cody-shared/src/editor'
 import { FeatureFlag, FeatureFlagProvider } from '@sourcegraph/cody-shared/src/experimentation/FeatureFlagProvider'
-import { annotateAttribution, Guardrails } from '@sourcegraph/cody-shared/src/guardrails'
 import { Result } from '@sourcegraph/cody-shared/src/local-context'
 import { MAX_BYTES_PER_FILE, NUM_CODE_RESULTS, NUM_TEXT_RESULTS } from '@sourcegraph/cody-shared/src/prompt/constants'
 import { truncateTextNearestLine } from '@sourcegraph/cody-shared/src/prompt/truncation'
@@ -64,8 +64,7 @@ import {
 } from './chat-helpers'
 import { ChatHistoryManager } from './ChatHistoryManager'
 import { addWebviewViewHTML, CodyChatPanelViewType } from './ChatManager'
-import { ChatViewProviderWebview } from './ChatPanelProvider'
-import { Config, IChatPanelProvider } from './ChatPanelsManager'
+import { ChatViewProviderWebview, Config, IChatPanelProvider } from './ChatPanelsManager'
 import { CodebaseStatusProvider } from './CodebaseStatusProvider'
 import { InitDoer } from './InitDoer'
 import { DefaultPrompter, IContextProvider, IPrompter } from './prompt'
@@ -76,7 +75,6 @@ interface SimpleChatPanelProviderOptions {
     config: Config
     extensionUri: vscode.Uri
     authProvider: AuthProvider
-    guardrails: Guardrails
     chatClient: ChatClient
     embeddingsClient: CachedRemoteEmbeddingsClient
     localEmbeddings: LocalEmbeddingsController | null
@@ -106,7 +104,6 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
 
     private config: Config
     private readonly authProvider: AuthProvider
-    private readonly guardrails: Guardrails
     private readonly chatClient: ChatClient
     private readonly embeddingsClient: CachedRemoteEmbeddingsClient
     private readonly codebaseStatusProvider: CodebaseStatusProvider
@@ -134,7 +131,6 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
         extensionUri,
         featureFlagProvider,
         authProvider,
-        guardrails,
         chatClient,
         embeddingsClient,
         localEmbeddings,
@@ -154,7 +150,6 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
         this.symf = symf
         this.editor = editor
         this.treeView = treeView
-        this.guardrails = guardrails
         this.recipeAdapter = recipeAdapter
         this.chatModel = new SimpleChatModel(this.selectModel(models))
         this.sessionID = this.chatModel.sessionID
@@ -487,11 +482,6 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
         this.postViewTranscript()
     }
 
-    public clearChatHistory(): Promise<void> {
-        // HACK: this is a no-op now. This exists only to satisfy the IChatPanelProvider interface
-        // and can be removed once we retire the old ChatPanelProvider
-        return Promise.resolve()
-    }
     public handleChatTitle(title: string): void {
         this.chatModel.setChatTitle(title)
         if (this.webviewPanel) {
@@ -560,7 +550,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
                 await vscode.commands.executeCommand('cody.settings.commands')
                 return
             }
-            return this.executeRecipe('custom-prompt', text.trim(), 'chat', userContextFiles, addEnhancedContext)
+            return this.executeRecipe('custom-prompt', text.trim(), 'chat', userContextFiles, false)
         }
 
         const displayText = userContextFiles?.length
@@ -584,10 +574,10 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
                 promptText: authStatus.endpoint && isDotCom(authStatus.endpoint) ? text : undefined,
                 contextSummary,
             }
-            telemetryService.log('CodyVSCodeExtension:recipe:chat-question:executed', properties, {
+            telemetryService.log('CodyVSCodeExtension:chat-question:recipe-used', properties, {
                 hasV2Event: true,
             })
-            telemetryRecorder.recordEvent('cody.recipe.chat-question', 'executed', {
+            telemetryRecorder.recordEvent('cody.recipe.chat-question', 'recipe-used', {
                 metadata: { ...contextSummary },
             })
         })
@@ -613,6 +603,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
             ...localProcess,
             debugEnable: config.debugEnable,
             serverEndpoint: config.serverEndpoint,
+            experimentalGuardrails: config.experimentalGuardrails,
         }
         await this.postMessage({ type: 'config', config: configForWebview, authStatus })
         logDebug('SimpleChatPanelProvider', 'updateViewConfig', { verbose: configForWebview })
@@ -687,7 +678,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
                     )
                 },
                 close: content => {
-                    this.addBotMessageWithGuardrails(requestID, content)
+                    this.addBotMessage(requestID, content)
                 },
                 error: (partialResponse, error) => {
                     if (isAbortError(error)) {
@@ -848,61 +839,29 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
     }
 
     /**
-     * Finalizes adding a bot message to the chat model, with guardrails, and triggers an
-     * update to the view.
+     * Finalizes adding a bot message to the chat model and triggers an update to the view.
      */
-    private addBotMessageWithGuardrails(requestID: string, rawResponse: string): void {
-        this.guardrailsAnnotateAttributions(reformatBotMessageForChat(rawResponse, ''))
-            .then(displayText => {
-                this.chatModel.addBotMessage({ text: rawResponse }, displayText)
-                void this.saveSession()
-                this.postViewTranscript()
+    private addBotMessage(requestID: string, rawResponse: string): void {
+        const displayText = reformatBotMessageForChat(rawResponse, '')
+        this.chatModel.addBotMessage({ text: rawResponse }, displayText)
+        void this.saveSession()
+        this.postViewTranscript()
 
-                // Count code generated from response
-                const codeCount = countGeneratedCode(rawResponse)
-                if (codeCount?.charCount) {
-                    // const metadata = lastInteraction?.getHumanMessage().metadata
-                    telemetryService.log(
-                        'CodyVSCodeExtension:chatResponse:hasCode',
-                        { ...codeCount, requestID },
-                        { hasV2Event: true }
-                    )
-                    telemetryRecorder.recordEvent('cody.chatResponse.new', 'hasCode', {
-                        metadata: {
-                            ...codeCount,
-                        },
-                    })
-                }
-            })
-            .catch(error => {
-                throw error
-            })
-    }
-
-    private async guardrailsAnnotateAttributions(text: string): Promise<string> {
-        if (!this.config.experimentalGuardrails) {
-            return text
-        }
-
-        const result = await annotateAttribution(this.guardrails, text)
-
-        // Only log telemetry if we did work (ie had to annotate something).
-        if (result.codeBlocks > 0) {
+        // Count code generated from response
+        const codeCount = countGeneratedCode(rawResponse)
+        if (codeCount?.charCount) {
+            // const metadata = lastInteraction?.getHumanMessage().metadata
             telemetryService.log(
-                'CodyVSCodeExtension:guardrails:annotate',
-                {
-                    codeBlocks: result.codeBlocks,
-                    duration: result.duration,
-                },
+                'CodyVSCodeExtension:chatResponse:hasCode',
+                { ...codeCount, requestID },
                 { hasV2Event: true }
             )
-            telemetryRecorder.recordEvent('cody.guardrails.annotate', 'executed', {
-                // Convert nanoseconds to milliseconds to match other telemetry.
-                metadata: { codeBlocks: result.codeBlocks, durationMs: result.duration / 1000000 },
+            telemetryRecorder.recordEvent('cody.chatResponse.new', 'hasCode', {
+                metadata: {
+                    ...codeCount,
+                },
             })
         }
-
-        return result.text
     }
 
     public setConfiguration(newConfig: Config): void {
@@ -958,7 +917,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
                     )
                 },
                 close: (responseText: string) => {
-                    this.addBotMessageWithGuardrails(requestID, responseText)
+                    this.addBotMessage(requestID, responseText)
                 },
                 error: (partialResponse: string, error: Error) => {
                     if (isAbortError(error)) {
@@ -1062,7 +1021,7 @@ class ContextProvider implements IContextProvider {
 
     private getVisibleEditorContext(): ContextItem[] {
         const visible = this.editor.getActiveTextEditorVisibleContent()
-        if (!visible) {
+        if (!visible || isCodyIgnoredFile(vscode.Uri.file(visible.fileName))) {
             return []
         }
         return [
@@ -1075,8 +1034,6 @@ class ContextProvider implements IContextProvider {
 
     public async getEnhancedContext(text: string): Promise<ContextItem[]> {
         const searchContext: ContextItem[] = []
-        let localEmbeddingsError
-        let remoteEmbeddingsError
         logDebug('SimpleChatPanelProvider', 'getEnhancedContext > embeddings (start)')
         let hasEmbeddingsContext = false
         const localEmbeddingsResults = this.searchEmbeddingsLocal(text)
@@ -1087,7 +1044,6 @@ class ContextProvider implements IContextProvider {
             searchContext.push(...r)
         } catch (error) {
             logDebug('SimpleChatPanelProvider', 'getEnhancedContext > local embeddings', error)
-            localEmbeddingsError = error
         }
         try {
             const r = await remoteEmbeddingsResults
@@ -1095,16 +1051,8 @@ class ContextProvider implements IContextProvider {
             searchContext.push(...r)
         } catch (error) {
             logDebug('SimpleChatPanelProvider', 'getEnhancedContext > remote embeddings', error)
-            remoteEmbeddingsError = error
         }
         logDebug('SimpleChatPanelProvider', 'getEnhancedContext > embeddings (end)')
-        if (localEmbeddingsError && remoteEmbeddingsError) {
-            throw new Error(
-                `local and remote embeddings search failed (local: ${getErrorMessage(
-                    localEmbeddingsError
-                )}) (remote: ${getErrorMessage(remoteEmbeddingsError)})`
-            )
-        }
 
         if (!hasEmbeddingsContext && this.symf) {
             try {
@@ -1162,6 +1110,9 @@ class ContextProvider implements IContextProvider {
         const r0 = (await this.symf.getResults(userText, [workspaceRoot])).flatMap(async results => {
             const items = (await results).flatMap(async (result: Result): Promise<ContextItem[] | ContextItem> => {
                 const uri = vscode.Uri.file(result.file)
+                if (isCodyIgnoredFile(uri)) {
+                    return []
+                }
                 const range = new vscode.Range(
                     result.range.startPoint.row,
                     result.range.startPoint.col,
@@ -1210,12 +1161,16 @@ class ContextProvider implements IContextProvider {
                 new vscode.Position(result.endLine, 0)
             )
             const uri = relativeFileUri('', result.fileName, range)
-            contextItems.push({
-                uri,
-                range,
-                text: result.content,
-                source: 'embeddings',
-            })
+
+            // Filter out ignored files
+            if (!isCodyIgnoredFile(vscode.Uri.file(result.fileName))) {
+                contextItems.push({
+                    uri,
+                    range,
+                    text: result.content,
+                    source: 'embeddings',
+                })
+            }
         }
         return contextItems
     }
@@ -1248,12 +1203,14 @@ class ContextProvider implements IContextProvider {
                 new vscode.Position(codeResult.startLine, 0),
                 new vscode.Position(codeResult.endLine, 0)
             )
-            contextItems.push({
-                uri,
-                range,
-                text: codeResult.content,
-                source: 'embeddings',
-            })
+            if (!isCodyIgnoredFile(vscode.Uri.file(path.join(codebase.local, codeResult.fileName)))) {
+                contextItems.push({
+                    uri,
+                    range,
+                    text: codeResult.content,
+                    source: 'embeddings',
+                })
+            }
         }
 
         for (const textResult of embeddings.textResults) {
@@ -1262,12 +1219,14 @@ class ContextProvider implements IContextProvider {
                 new vscode.Position(textResult.startLine, 0),
                 new vscode.Position(textResult.endLine, 0)
             )
-            contextItems.push({
-                uri,
-                range,
-                text: textResult.content,
-                source: 'embeddings',
-            })
+            if (!isCodyIgnoredFile(vscode.Uri.file(path.join(codebase.local, textResult.fileName)))) {
+                contextItems.push({
+                    uri,
+                    range,
+                    text: textResult.content,
+                    source: 'embeddings',
+                })
+            }
         }
 
         return contextItems
@@ -1337,7 +1296,7 @@ class ContextProvider implements IContextProvider {
         // global pattern for readme file
         const readmeGlobalPattern = '{README,README.,readme.,Readm.}*'
         const readmeUri = (await vscode.workspace.findFiles(readmeGlobalPattern, undefined, 1)).at(0)
-        if (!readmeUri) {
+        if (!readmeUri || isCodyIgnoredFile(readmeUri)) {
             return []
         }
         const readmeDoc = await vscode.workspace.openTextDocument(readmeUri)
@@ -1478,13 +1437,6 @@ function extractQuestion(input: string): string | undefined {
 
 function isAbortError(error: Error): boolean {
     return error.message === 'aborted' || error.message === 'socket hang up'
-}
-
-function getErrorMessage(error: unknown): string {
-    if (error instanceof Error) {
-        return error.message
-    }
-    return String(error)
 }
 
 function getContextWindowForModel(authStatus: AuthStatus, modelID: string): number {