+++ vscode/src/chat/local-code-search.ts
@@ -9,7 +9,7 @@ import { IndexedKeywordContextFetcher, Result } from '@sourcegraph/cody-shared/s
 import { MAX_HUMAN_INPUT_TOKENS } from '@sourcegraph/cody-shared/src/prompt/constants'
 import { truncateText } from '@sourcegraph/cody-shared/src/prompt/truncation'
 
-import { getActiveEditor } from '../editor/active-editor'
+import { getEditor } from '../editor/active-editor'
 
 /**
  * Uses symf to run an LLM-enhanced indexed keyword search for the user's query
@@ -164,7 +164,7 @@ ${firstNLines(text, 10)}
 }
 
 function getCurrentWorkspaceRoot(): string | null {
-    const uri = getActiveEditor()?.document?.uri
+    const uri = getEditor().active?.document?.uri
     if (uri) {
         const wsFolder = vscode.workspace.getWorkspaceFolder(uri)
         if (wsFolder) {